requirements.txt
flask==2.3.2
Bootstrap_Flask~=2.2.0
flask_sqlalchemy~=3.1.1
python-dotenv~=1.0.0
wtforms~=3.1.1
Flask-WTF~=1.2.1
Flask-Login~=0.6.3
Flask_CKEditor~=0.5.0
Werkzeug~=3.0.1
SQLALchemy~=2.0.23